Crate escrow_io

Structs

InitEscrow

Initializes an escrow program.

Wallet

Escrow wallet information.

Enums

EscrowAction

An enum to send the program info about what it should do.

EscrowEvent

An enum that contains a result of processed EscrowAction.

EscrowState

An enum for requesting the program state.

EscrowStateReply

An enum that contains a reply for a requested EscrowState.

WalletState

An escrow wallet state.

Type Definitions

WalletId

An escrow wallet ID.